+++ Logger.py
@@ -1,0 +1,6 @@
+import sys
+
+
+def log(prefix: str, message, *arg):
+    print(prefix+':', message, *arg)
+    sys.stdout.flush()
