+++ Voice.py
@@ -35,6 +35,7 @@ class Voice:
         self.t = kthread.KThread(target=self.voice_exec, name="Voice", daemon=True)
         self.t.start()
         self.v_id = 1
+        self.is_playing = False
 
     def say(self, vSay):
         if self.v_enabled:
@@ -71,8 +72,10 @@ class Voice:
                 words = self.q.get(timeout=1)
                 self.q.task_done()
                 if words is not None:
+                    self.is_playing = True
                     engine.say(words)
                     engine.runAndWait()
+                    self.is_playing = False
             except:
                 pass
